Moving FuncEnv into a separate file
