+++ crates/winch/src/compiler.rs
@@ -3,44 +3,15 @@ use cranelift_codegen::{Final, MachBufferFinalized};
 use object::write::{Object, SymbolId};
 use std::any::Any;
 use std::sync::Mutex;
-use wasmparser::{FuncValidatorAllocations, types::Types};
+use wasmparser::FuncValidatorAllocations;
 use wasmtime_cranelift_shared::obj::ModuleTextBuilder;
 use wasmtime_environ::{
     CompileError, DefinedFuncIndex, FilePos, FuncIndex, FunctionBodyData, FunctionLoc,
-    ModuleTranslation, ModuleTypes, PrimaryMap, Tunables, WasmFunctionInfo, Module,
+    ModuleTranslation, ModuleTypes, PrimaryMap, Tunables, WasmFunctionInfo,
 };
-use winch_codegen::{TargetIsa, Callee};
+use winch_codegen::TargetIsa;
 
-// Copied from winch/environ/src/lib.rs, REPLACE
-pub struct FuncEnv<'a> {
-    /// The translated WebAssembly module.
-    pub module: &'a Module,
-    /// Type information about a module, once it has been validated.
-    pub types: &'a Types,
-}
-
-impl<'a> winch_codegen::FuncEnv for FuncEnv<'a> {
-    fn callee_from_index(&self, index: u32) -> Callee {
-        let func = self
-            .types
-            .function_at(index)
-            .unwrap_or_else(|| panic!("function type at index: {}", index));
-
-        Callee {
-            ty: func.clone(),
-            import: self.module.is_imported_function(FuncIndex::from_u32(index)),
-            index,
-        }
-    }
-}
-
-impl<'a> FuncEnv<'a> {
-    /// Create a new function environment.
-    pub fn new(module: &'a Module, types: &'a Types) -> Self {
-        Self { module, types }
-    }
-}
-
+use crate::environ::FuncEnv;
 
 pub(crate) struct Compiler {
     isa: Box<dyn TargetIsa>,
