+++ crates/winch/src/lib.rs
@@ -2,3 +2,4 @@ pub use builder::builder;
 
 mod builder;
 mod compiler;
+mod environ;
